refactor: separate platform utils
