+++ platforms/evm/src/chain.ts
@@ -1,7 +1,4 @@
-import {
-  ChainContext,
-  Platform,
-} from '@wormhole-foundation/connect-sdk';
+import { ChainContext, Platform } from '@wormhole-foundation/connect-sdk';
 import { EvmPlatform } from './platform';
 
 export class EvmChain extends ChainContext<'Evm'> {
