+++ platforms/evm/src/protocols/circleBridge.ts
@@ -11,9 +11,7 @@ import {
   Network,
 } from '@wormhole-foundation/connect-sdk';
 
-import {
-  evmNetworkChainToEvmChainId,
-} from '../constants';
+import { evmNetworkChainToEvmChainId } from '../constants';
 import {
   addFrom,
   addChainId,
